# fix issue 7054 - format() aligns using code units instead of graphemes