I didn't see it in bugzilla before cooking the fix but apparently it's controversial for some people.