@biozic 2018-01-07
Yes, having a formatted string that is longer than what the width specification specifies for some inputs is probably controversial. The workaround is to transcode the input to `dstring`. That's why I closed my bugzilla issue.