@BBasile 2018-01-07
To me it's more logic to get a well aligned output, whatever is the encoding. What do the foreign standard libraries do ?