@biozic 2018-01-07
Python 3:
```python
print('|{:>8}|'.format('abc'))
print('|{:>8}|'.format('αβγ'))
```
prints
```
|     abc|
|     αβγ|
```