@BBasile 2018-01-07
[Go too](https://play.golang.org/p/xnvXrubWsCl) so your report was valid after all. I don't say that we should align on the others but `format()` is one quite a universal function.