@JackStouffer 2018-01-07
This doesn't really fix the problem:

```d
auto a = "\u0065\u0301";
writeln(a); // é
assert(a.walklength == 2);
```

Operating on the code point level rather than the code unit level just opens up different different issues.

If you want the same behavior as Python, you need to use `byGrapheme.walkLength`, but it's much slower. I think there is a case to be made here and would be much more sensible to the user in a string formatting function.

There's also the issue of backwards incompatibility.