@JackStouffer 2018-01-07
I would changes this to use `byGrapheme` and solicit opinions from other Phobos maintainers.