@biozic 2018-01-07
Yes `byGrapheme.walkLength` is more to the point, since the problem was that the number of graphemes was inconsistent.

If performance is an issue, we could add a specific flag, say `!` in the format-specification grammar, to tell that the width specification is not a number of code units, but a number of graphemes, e.g. `%8s` for the former, `%!8s` for the latter.