@BBasile 2018-01-07
`byGrapheme.walkLength` is not CTFE-able but this seems to be a requirement for `format()`.