@BBasile 2018-01-07
> Operating on the code point level rather than the code unit level just opens up different issues.

_different_ issues ? It seems that what would happen is that in the current state, the fix would be incomplete, e.g only working when `walkLength == byGrapheme.walkLength`.

> There's also the issue of backwards incompatibility.

Likely only some ugly workarounds, which then would not be necessary anymore.
The most simple workaround that people could used (calling  `spec.format(stuff.to!dstring)`) would still work.