@JackStouffer 2018-01-08
It brings up different issues because it's slower and it breaks from an easy to understand but simplistic rule of string width == length, while also not changing the output if the string was made with or normalized with decomposing rules. With either code points or code units, the behavior is not intuitive but this method gives the *illusion* of correctness.

I don't have the time atm to throughly investigate this but I don't off the bat see any reason why `byGrapheme` wound't work in CTFE, especially since encoding/decoding work at CTFE. Fixing this would be the right way to go.