@BBasile 2018-01-08
@JackStouffer, just compiles this

```d
import std.range, std.uni;
enum len = "abc".byGrapheme.walkLength;
```

To see what i'm talking about. If you think that `byGrapheme` is CTFE-able maybe that's because something is not **yet** tested.