@quickfur 2018-01-08
This is a long-standing issue. See: https://issues.dlang.org/show_bug.cgi?id=7054

It's not so easy to fix because of safety, purity, and performance issues with byGrapheme. But if you can make it work, it will be awesome.