@quickfur 2018-01-08
@JackStouffer  The last time I checked, byGrapheme didn't work in CTFE because for runtime performance reasons it uses some hacks to store the first few characters in situ, so that memory allocation won't be needed unless it's an unusually-long grapheme cluster.  But there's probably a way to make it work by special-casing it for CTFE.