@JackStouffer 2018-01-08
@quickfur Are you aware of a method of using static if to check if code is being executed in CTFE? If that's possible, then this is an easy fix.