@BBasile 2018-01-08
I'm a partisan of _"this is still better than nothing"_, meaning:
- merge this partial fix
- open a proper issue on bugzilla: _byGrapheme not usable with CTFE_
- document the issue that's still there in the code. If someone wants to fix it he'll likely reach the comment saying _see http..._

About the if `(__ctfe)`. It appears that invalid branches are eliminated by the optimizer but in reality the two branches have to be valid before. So, i guess i teach you nothing, this is not usable for now, it's not like a `static if`.

Finally i'd like to say that it's a shame that Phobos cannot be right for this format thing "just" because of a core feature that is CTFE. Paradoxically.