@JackStouffer 2018-01-08
Another possibility is to use this functionality for CTFE and byGrapheme for runtime.