@quickfur 2018-01-08
The problem is that ever since `std.format` because CTFE-able, lots of code have come to rely on it, for better or for worse.  So we can't afford to break that now.

So here's a possible approach that allows us to at least make progress on *one* front, without holding everything up until somebody figures out how to make `byGrapheme` CTFE-able:

1. Use `if (__ctfe)` to isolate the current wrong width formatting to CTFE only, and use the `!__ctfe` branch to implement walking by grapheme cluster length.

2. File a bug to the effect that width formatting is wrong in CTFE.

3. Merge this fix, so that we get the correct behaviour at least at runtime.

4. Work on making byGrapheme CTFE-able, then fix the rest of the bug.