@JackStouffer 2018-01-08
So for this PR, I'd approve it if it used walklength in CTFE, used byGrapheme in runtime, added a note to the docs about this discrepancy, and added a changelog entry.