@quickfur 2018-01-08
@JackStouffer Unfortunately, we can't mix `static if` with CTFE. See: https://wiki.dlang.org/User:Quickfur/Compile-time_vs._compile-time