@BBasile 2018-01-08
That's the idea assuming the difference between CT and RT can be made.