@quickfur 2018-01-08
I just checked, it *is* possible to make this partial fix work. The trick here is that `byGrapheme` is compilable, and code that depends on it can be used in CTFE as long as the CTFE engine never actually evaluates it. So here's a proof of concept:

````
import std.range.primitives;
import std.stdio;
import std.uni;

size_t width(string s) {
        if (__ctfe) {
                return s.length;
        } else {
                return s.byGrapheme.walkLength;
        }
}

enum str = "Это досадно!";
pragma(msg, width(str));
void main() {
        writeln(width(str));
}
````

Here's the compiler output:
````
22LU
````

And here's the runtime output:
````
12
````

This means that a simple `if (__ctfe)` branch can be used, for now, to implement the partial fix.  Let's do that first, then we can think about how to make `byGrapheme` itself work in CTFE.