@BBasile 2018-01-08
Good. I thought i wouldn't work. Although it's still strange to have different result at CT and RT, so just `walkLength` for now ?