@BBasile 2018-01-09
The onlyAscii optimization allows to maintain current performances but new benchmarks are worst than announced previously: 4  to 7  times slower. You'll have to choose between speed or correctness.

```d
#!dmd -O -release -inline -boundscheck=off
import std.stdio, std.datetime.stopwatch, std.format;

string test_ascii = "abcdefghijkl";
string test_utf_n = "ÃŸÃ©Ã Ã§ÃŸÃ©Ã Ã§ÃŸÃ©Ã Ã§";
string test_utf_g = "e\u0301e\u0301e\u0301e\u0301e\u0301e\u0301";

/*
before:
46 ms, 904 Î¼s, and 2 hnsecs
23 ms, 466 Î¼s, and 9 hnsecs
36 ms, 789 Î¼s, and 2 hnsecs

after, w/o ascii opt:
136 ms, 802 Î¼s, and 7 hnsecs
150 ms, 675 Î¼s, and 5 hnsecs
121 ms, 825 Î¼s, and 9 hnsecs

after, w/ ascii opt:
49 ms and 769 Î¼s
148 ms, 805 Î¼s, and 5 hnsecs
120 ms, 711 Î¼s, and 5 hnsecs
*/

void main()
{
    benchmark!((){auto s = format("%20s", test_ascii);})(100_000)[0].writeln;
    benchmark!((){auto s = format("%20s", test_utf_n);})(100_000)[0].writeln;
    benchmark!((){auto s = format("%20s", test_utf_g);})(100_000)[0].writeln;
}
```